First-off, found out that /html/cl4 directory was empty.  Copied over existing cl4 stuff from SVN template site, and styles and javascript behaviour returned as expected.
Set view_flag to false for all visible fields in DemoSub model.  The missing part of this issue is likely that there needs to be a warning that there are no visible fields for this model.

Added an $_any_visible member to ORM to track if any fields are visible.  Set during prepare_form().  Cleaned up the phpdocs for ORM's members.

Passing $_any_visible along to the view for displaying the view of a model.  If not set, just displaying the text "No fields in this model are visible."  Did the same for editing an item.

Augh, again we have to deal with MuliORM not being able to know the internals of its model's ORM stuff, so I had to create an any_visible() function in ORM for MultiORM to use.  Using that to set $any_visible inside the orm_editable_list view.